+++ common/src/main/java/net/spell_engine/api/config/ArmorSetConfig.java
@@ -1,5 +1,7 @@
 package net.spell_engine.api.config;
 
+import net.fabricmc.loader.api.FabricLoader;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,12 +19,20 @@ public class ArmorSetConfig {
     public static class Piece {
         public Piece() {
         }
+        public Piece(int armor) {
+            this.armor = armor;
+        }
 
         public int armor = 0;
         public ArrayList<AttributeModifier> attributes = new ArrayList<>();
-
-        public Piece(int armor) {
-            this.armor = armor;
+        public ConditionalAttributes conditional_attributes = null;
+        public List<AttributeModifier> selectedAttributes() {
+            if (this.conditional_attributes != null
+                    && this.conditional_attributes.required_mod() != null
+                    && FabricLoader.getInstance().isModLoaded(this.conditional_attributes.required_mod())) {
+                return this.conditional_attributes.attributes();
+            }
+            return this.attributes;
         }
 
         public Piece add(AttributeModifier attribute) {
@@ -32,6 +42,11 @@ public class ArmorSetConfig {
 
         public Piece addAll(List<AttributeModifier> attributes) {
             this.attributes.addAll(attributes);
+            return this;
+        }
+
+        public Piece addConditional(String required_mod, List<AttributeModifier> attributes) {
+            this.conditional_attributes = new ConditionalAttributes(required_mod, attributes);
             return this;
         }
     }