+++ common/src/main/java/net/spell_engine/api/config/ConditionalAttributes.java
@@ -1,0 +1,6 @@
+package net.spell_engine.api.config;
+
+import java.util.List;
+
+public record ConditionalAttributes(String required_mod, List<AttributeModifier> attributes) {
+}